+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Subscribing to /r/all ([#28](https://github.com/Tunous/Dank/pull/28))
 - View parent submission of crosspost ([#29](https://github.com/Tunous/Dank/pull/29))
 - Links to repository and changelog in about screen ([#42](https://github.com/Tunous/Dank/pull/42))
+- Colored comments depth level ([#49](https://github.com/Tunous/Dank/pull/49))
 
 ### Fixed
 - Clicking on notification doesn't open downloaded image ([#35](https://github.com/Tunous/Dank/pull/35))